Proper setup file.
